app/main.py
import os
from flask import Flask, jsonify
import psycopg2

app = Flask(__name__)


def get_db_connection():
    database_url = os.environ.get("DATABASE_URL")
    if not database_url:
        raise RuntimeError("DATABASE_URL environment variable is not set")
    conn = psycopg2.connect(database_url)
    return conn


@app.route("/health")
def health():
    return jsonify({"status": "ok"})


@app.route("/users")
def get_users():
    conn = get_db_connection()
    cur = conn.cursor()
    cur.execute("SELECT id, name, email FROM users;")
    rows = cur.fetchall()
    cur.close()
    conn.close()
    users = [{"id": r[0], "name": r[1], "email": r[2]} for r in rows]
    return jsonify(users)


if __name__ == "__main__":
    app.run(host="0.0.0.0", port=5000)
